Experimental: added 64-bit pos, mate pos and insert size to CRAM V4.0.
The BAM internal layout is 64-bit too, and SAM can fill this out with
64-bit fields.  This permits SAM to CRAM to SAM round trips with long
references, although BAM is still unable to handle this (and it
doesn't report this failure yet).

+++ io_lib/cram_stats.h
@@ -45,8 +45,8 @@ extern "C" {
 #include "io_lib/hash_table.h"
 
 cram_stats *cram_stats_create(void);
-void cram_stats_add(cram_stats *st, int32_t val);
-void cram_stats_del(cram_stats *st, int32_t val);
+void cram_stats_add(cram_stats *st, int64_t val);
+void cram_stats_del(cram_stats *st, int64_t val);
 void cram_stats_dump(cram_stats *st);
 void cram_stats_free(cram_stats *st);
